discord-bot-gilberto/pyproject.toml
[project]
name = "discord-bot-gilberto"
version = "0.1.0"
description = "Discord bot for managing commands for a specific channel"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "allure-pytest>=2.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]